Move functionality to the library for clear interface
