+++ routesrv/eskipbytes.go
@@ -1,4 +1,4 @@
-package main
+package routesrv
 
 import (
 	"bytes"
