+++ routesrv/options.go
@@ -1,4 +1,4 @@
-package options
+package routesrv
 
 import (
 	"regexp"
